refactor(sdk): move DefinitionError to errors namespace
