+++ packages/sdk/src/bot/definition.ts
@@ -5,7 +5,7 @@ import { IntegrationPackage, PluginPackage } from '../package'
 import { PluginInterfaceExtension, PluginIntegrationExtension } from '../plugin'
 import { SchemaDefinition } from '../schema'
 import * as utils from '../utils'
-import { DefinitionError } from '../utils/error-utils'
+import { DefinitionError } from '../errors'
 import { ValueOf, Writable, Merge, StringKeys } from '../utils/type-utils'
 import { SDK_VERSION } from '../version'
 import { z } from '../zui'
