+++ packages/sdk/src/integration/definition/index.ts
@@ -3,7 +3,7 @@ import { SchemaTransformOptions } from '../../common/types'
 import { resolveInterface } from '../../interface/resolve'
 import { InterfacePackage } from '../../package'
 import * as utils from '../../utils'
-import { DefinitionError } from '../../utils/error-utils'
+import { DefinitionError } from '../../errors'
 import { SDK_VERSION } from '../../version'
 import { z } from '../../zui'
 import { SchemaStore, BrandedSchema, createStore, isBranded, getName } from './branded-schema'
